Return HTTP error when ui_update_acls fails
This will cause the nightly cron job to fail, which will trigger an error
email, which will make it obvious when there's a problem. If we don't do
this, we just get no email, which makes it hard to notice the issue unless
someone remembers to actively look out for the email.