Enhance update-running-status determination logic
Currently, if there's an issue determining whether an update is running,
the code "fails safe" and assumes that it is. While this is safe, it's
problematic since any problems are not self-clearing. Enhance the code to
assume that the update is _not_ running some time after it was triggered,
so that any problems auto-clear.